fix delete, show image in student update

+++ php/deleteAnnouncement.php
@@ -11,8 +11,10 @@
 		$stmt->execute();
 		
 		foreach(($stmt->fetchAll()) as $row) { 
+			if($row['file']<>'') {
 			$myFile = $_SERVER["DOCUMENT_ROOT"] . "/uploads/" . $row['file'];
 			unlink($myFile) or die("Couldn't delete file.");
+			}
 		}
 		
 		$stmt = $conn->prepare("DELETE FROM posts WHERE id = :id");
